requirements.txt
django
djangorestframework
reportlab
django-jazzmin